@@ -16,16 +16,16 @@
     "@types/node": "^22.10.2",
     "autoprefixer": "^10.4.20",
     "postcss": "^8.4.49",
-    "svelte": "^4.2.19",
-    "svelte-check": "^3.8.6",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tailwindcss": "^3.4.17",
     "tslib": "^2.8.1",
     "typescript": "^5.7.2",
-    "vite": "^5.4.11"
+    "vite": "^6.3.0"
   },
   "dependencies": {
     "@auth/sveltekit": "^1.5.2",
-    "@lucide/svelte": "^0.469.0",
+    "@lucide/svelte": "^0.575.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "tailwind-merge": "^2.6.0",
